weco-backend/requirements.txt
# Backend зависимости для WEco
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-decouple==3.8
qrcode[pil]==7.4.2
Pillow==10.1.0
httpx==0.25.2
bcrypt==3.2.2